% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{aggregateXYZs}
\alias{aggregateXYZs}
\title{Aggregate xyz values by xyz group}
\usage{
aggregateXYZs(xyz.list, FUN = mean, na.rm = TRUE, ...)
}
\arguments{
\item{xyz.list}{A list of list with xyz values.}

\item{FUN}{The function for aggregate the values.}

\item{na.rm}{Remove the NA values or not. If TRUE, NA values will not be 
considered. Otherwise, the NA values will be filled with the mean of
nearby values.}

\item{...}{Other parameters for the FUN.}
}
\value{
A list of xyz.
}
\description{
Aggregate the corresponding xyz values of xyz values for each xyz group
by given function.
}
