% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{getFeatureGR}
\alias{getFeatureGR}
\title{Create annotation features}
\usage{
getFeatureGR(
  txdb,
  org,
  range,
  keytype = "ENTREZID",
  geneSymbolColumn = "SYMBOL",
  cols = seq.int(7)
)
}
\arguments{
\item{txdb}{An \link[GenomicFeatures:TxDb-class]{TxDb} object.}

\item{org}{An \link[AnnotationDbi:AnnotationDb-class]{OrgDb} object.}

\item{range}{\link[GenomicRanges:GRanges-class]{GRanges} object. The coordinates.}

\item{geneSymbolColumn, keytype}{The column names in the OrgDb for key of TxDb
and the gene symnbols.}

\item{cols}{The colors for each gene.}
}
\value{
A \link[GenomicRanges:GRanges-class]{GRanges} object
}
\description{
Generate annotation features (feature.gr for \link[geomeTriD]{view3dStructure})
from TxDb and Org object.
}
\examples{
# example code

}
