% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{getMedoidId}
\alias{getMedoidId}
\title{Find the cluster medoid sample}
\usage{
getMedoidId(dist, cluster, N = 1L, na.rm = TRUE)
}
\arguments{
\item{dist}{A dist object}

\item{cluster}{The group information for the elements in the dist.}

\item{N}{The number of medoid sample to be returned.}

\item{na.rm}{A logical evaluating to TRUE or FALSE indicating whether NA 
values should be stripped before the computation proceeds.}
}
\value{
The medoid element labels for each cluster.
}
\description{
Find the medoid sample of each cluster in a given distance matrix
}
\examples{
x <- matrix(rnorm(100), nrow = 5)
dist <- dist(x)
group <- c('A', 'A', 'B', 'B', 'B')
getMedoidId(dist, group)
}
