% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{import3dg}
\alias{import3dg}
\title{Import Hickit output}
\usage{
import3dg(
  filenames,
  comment.char = "#",
  ...,
  parental_postfix = c("(pat)", "(mat)")
)
}
\arguments{
\item{filenames}{Vector of character. The 3dg file names of output of Hickit}

\item{comment.char, ...}{parameters pass to \link[utils:read.table]{rea.delim}.}

\item{parental_postfix}{the postfix of chromosome names. Default is 
c("(pat)", "(mat)"). 
For hickit, it maybe c('a', 'b').}
}
\value{
An object of GRangesList
}
\description{
Import 3dg file from Hickit output to a GRangesList Object
}
\examples{
f3dg <- system.file('extdata', 'GSE162511',
 'GSM4382149_cortex-p001-cb_001.20k.1.clean.3dg.txt.gz',
  package='geomeTriD.documentation')
xyz <- import3dg(f3dg)
}
