% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{importFLAMINGO}
\alias{importFLAMINGO}
\title{Import FLAMINGO output}
\usage{
importFLAMINGO(filenames)
}
\arguments{
\item{filenames}{Vector of character. The file names of output of FLAMINGO.}
}
\value{
An object of GRangesList
}
\description{
Import FLAMINGO output to a GRangesList Object
}
\examples{
chrs <- paste0('chr', c(1, 'X'))
gm12878 <- paste0('https://github.com/wangjr03/FLAMINGO/',
                  'raw/refs/heads/main/predictions/GM12878/',
                  chrs, '_5kb.txt')
names(gm12878) <- chrs
gm12878.gl <- importFLAMINGO(gm12878)

}
