% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{importGInteractionsFromUrl}
\alias{importGInteractionsFromUrl}
\title{Import genomic interactions from URL}
\usage{
importGInteractionsFromUrl(
  urls,
  resolution,
  range,
  format = "cool",
  normalization = "balanced"
)
}
\arguments{
\item{urls}{Vector of character. URLs.}

\item{resolution}{The resolution of the matrix.}

\item{range}{GRanges object. The coordinates.}

\item{format}{The format of the file.}

\item{normalization}{Normalization matrix in the file.}
}
\value{
A list of GInteractions object
}
\description{
Download files and import interactions
}
\examples{
# example code

}
