% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{importSignalFromUrl}
\alias{importSignalFromUrl}
\title{Import genomic signals from URL}
\usage{
importSignalFromUrl(urls, range, cols, format = "BigWig")
}
\arguments{
\item{urls}{Vector of character. URLs.}

\item{range}{\link[GenomicRanges:GRanges-class]{GRanges} object. The coordinates.}

\item{cols}{The colors for each signal.}

\item{format}{The format of the file.}
}
\value{
A list of track object
}
\description{
Download files and import signals.
}
\examples{
library(GenomicRanges)
range <- GRanges('chr4:50000-100000')
## set the URLs
prefix <- 'https://www.encodeproject.org/files/'
url <- c(H3K9me3 = 'ENCFF776OVW/@download/ENCFF776OVW.bigWig')
urls <- paste0(prefix, url)
names(urls) <- names(url)
# download the files and import the signals
# genomicSigs <- importSignalFromUrl(urls, range = range,
#                                     cols = list(
#                                       'H3K9me3'=c('#000011', 'blue'),
#                                       'H3K27ac'=c('#111100', 'orange')),
#                                     format = 'BigWig')

}
