% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{importTensorFLAMINGO}
\alias{importTensorFLAMINGO}
\title{Import Tensor-FLAMINGO output}
\usage{
importTensorFLAMINGO(filenames, binsize, chr)
}
\arguments{
\item{filenames}{Vector of character. The file names of output of Tensor-FLAMINGO.}

\item{binsize}{Bin size of the interaction. It will be the width of 
GRanges.}

\item{chr}{Chromosome name}
}
\value{
An object of GRangesList
}
\description{
Import Tensor-FLAMINGO output to a GRangesList Object
}
\examples{
# example code

}
