% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{paddingGRangesList}
\alias{paddingGRangesList}
\title{Padding GRanges List}
\usage{
paddingGRangesList(xyzs)
}
\arguments{
\item{xyzs}{A list of GRanges.}
}
\value{
A list with element 'gr' represent the genomic ranges and
'xyzs' represent the x,y,z positions in a data.frame.
}
\description{
Make all GRanges object in a list same length by filling with NA
}
\examples{
library(GenomicRanges)
gr <- GRanges('chr1', IRanges(1:10, width=1), x=1:10, y=1:10, z=1:10)
xyzs <- list(gr1=gr[c(1, 3, 5, 7, 9)],
             gr2=gr[c(2, 4, 6, 8, 10)])
paddingGRangesList(xyzs)

}
