% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{importFLAMINGO}
\alias{importFLAMINGO}
\title{Import FLAMINGO output}
\usage{
importFLAMINGO(filenames)
}
\arguments{
\item{filenames}{Vector of character. The file names of output of FLAMINGO.}
}
\value{
An object of GRangesList
}
\description{
Import FLAMINGO output to a GRangesList Object
}
\examples{
# example code

}
