% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{importGenomicSigs}
\alias{importGenomicSigs}
\title{Import genomic signals}
\usage{
importGenomicSigs(paths, range, cols, format = "BigWig")
}
\arguments{
\item{paths}{Vector of character. File path}

\item{range}{\link[GenomicRanges:GRanges-class]{GRanges} object. The coordinates.}

\item{cols}{The colors for each signal.}

\item{format}{The format of the file.}
}
\value{
A list of track object
}
\description{
Import genomic signals for given range.
}
\examples{
# example code

}
