% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{importGInteractionsFromUrl}
\alias{importGInteractionsFromUrl}
\title{Import genomic interactions from URL}
\usage{
importGInteractionsFromUrl(
  urls,
  resolution,
  range,
  format = "cool",
  normalization = "balanced"
)
}
\arguments{
\item{urls}{Vector of character. URLs.}

\item{resolution}{The resolution of the matrix.}

\item{range}{\link[GenomicRanges:GRanges-class]{GRanges} object. The coordinates.}

\item{format}{The format of the file.}

\item{normalization}{Normalization matrix in the file.}
}
\value{
A list of GInteractions object
}
\description{
Download files and import interactions
}
\examples{
library(GenomicRanges)
range_chr8<- GRanges('chr8:85550000-85800000')
url <- paste0("https://ftp.ncbi.nlm.nih.gov/geo/samples/GSM6281nnn/",
 "GSM6281851/suppl/GSM6281851_RCMC_BR1_merged_allCap_DMSO_mm39.merged.50.mcool")
gi <- importGInteractionsFromUrl(urls=url, resolution=500, range=range_chr8,
                                 format='cool', normalization='balanced')

}
