% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{importSignalFromUrl}
\alias{importSignalFromUrl}
\title{Import genomic signals from URL}
\usage{
importSignalFromUrl(urls, range, cols, format = "BigWig")
}
\arguments{
\item{urls}{Vector of character. URLs.}

\item{range}{\link[GenomicRanges:GRanges-class]{GRanges} object. The coordinates.}

\item{cols}{The colors for each signal.}

\item{format}{The format of the file.}
}
\value{
A list of track object
}
\description{
Download files and import signals.
}
\examples{
# example code

}
