% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{importSuperRec}
\alias{importSuperRec}
\title{Import SuperRec output}
\usage{
importSuperRec(filenames, superRecInputFilenames, binsize, chr)
}
\arguments{
\item{filenames}{Vector of character. The file names of output of SuperRec}

\item{superRecInputFilenames}{Vector of character.
The filenames of input of SuperRec.}

\item{binsize}{Bin size of the interaction. It will be the width of 
GRanges.}

\item{chr}{Chromosome name}
}
\value{
An object of GRangesList
}
\description{
Import SuperRec output to a GRangesList Object
}
\examples{
# example code

}
