% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{paddingGRangesList}
\alias{paddingGRangesList}
\title{Padding GRanges List}
\usage{
paddingGRangesList(xyzs)
}
\arguments{
\item{xyzs}{A list of GRanges.}
}
\value{
A list with element 'gr' represent the genomic ranges and
'xyzs' represent the x,y,z positions in a data.frame.
}
\description{
Make all GRanges object in a list same length by filling with NA
}
