% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{showPairs}
\alias{showPairs}
\title{Show a pair of geometries}
\usage{
showPairs(a, b, height = "95vh", ...)
}
\arguments{
\item{a, b}{A list of \link[geomeTriD:threeJsGeometry-class]{threeJsGeometry} object.}

\item{height}{The height of the widgets.}

\item{...}{The parameter for \link[geomeTriD:threeJsViewer]{threeJsViewer}.}
}
\value{
A \link[GenomicRanges:GRanges-class]{GRanges} object
}
\description{
Show a pair of geometries side by side.
}
\examples{
# example code

}
